+++ lib/ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: cchameyr <cchameyr@students.42.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/23 15:20:20 by cchameyr          #+#    #+#             */
-/*   Updated: 2015/11/25 23:32:45 by cchameyr         ###   ########.fr       */
+/*   Updated: 2015/11/29 10:03:56 by cchameyr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dst, const char *src, size_t n)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (src[i] && i < n)
